+++ mobile/src/libs/components/toast/toast.tsx
@@ -5,7 +5,7 @@ import LibraryToast, {
   type ToastConfigParams,
 } from 'react-native-toast-message';
 
-import { type NotificationType } from '#libs/enums/enums';
+import { type NotificationType } from '#libs/packages/notification/notification';
 import { type ValueOf } from '#libs/types/types';
 
 import { styles } from './styles';
