+++ mobile/src/libs/enums/enums.ts
@@ -1,6 +1,5 @@
 export { DataStatus } from './app/app';
 export { MainScreenName, RootScreenName } from './navigation/navigation';
-export { NotificationType } from './notification/notification';
 export { AppColor } from './ui/ui';
 export {
   APIPath,
