+++ mobile/src/libs/packages/notification/notification.ts
@@ -2,5 +2,9 @@ import { Notification } from './notification.package';
 
 const notification = new Notification();
 
-export { NotificationType } from '#libs/enums/enums';
+export { NotificationType } from './libs/enums/enums';
+export {
+  type NotificationPayload,
+  type NotificationProperties,
+} from './libs/types/types';
 export { notification };
