+++ mobile/src/libs/packages/store/middlewares/middlewares.ts
@@ -5,11 +5,11 @@ import {
   type Middleware,
 } from '@reduxjs/toolkit';
 
-import { NotificationType } from '#libs/enums/enums';
+import { NotificationType } from '#libs/packages/notification/notification';
 import { type AsyncThunkConfig } from '#libs/types/types';
 import { appActions } from '#slices/app/notifications';
 
-import { type AppDispatch } from '../libs/types/types';
+import { type AppDispatch } from '../store';
 
 const handleError: Middleware<AsyncThunkConfig, unknown, AppDispatch> = ({
   dispatch,
