+++ mobile/src/libs/packages/store/store.ts
@@ -4,3 +4,4 @@ import { Store } from './store.package';
 const store = new Store(config);
 
 export { store };
+export { type AppDispatch } from './libs/types/types';
